fix: resolve modules if folder contains a package.json file
getMappedPath only resolves to files that exist. This does not
work when a TSConfig path alias resolves to a module folder
containing built assets, and described by a package.json, e.g.
foo/dist.

If the tsconfig path is mapped to a folder, then getMappedPath
returns undefined as the folder is not a file. However, in somes
cases it is ok for this to resolve as the package.json contains
a `typings` entry that references a .d.ts file, which is
correctly handled by enhanced-resolve